DRY: Allow Rails to find ValidatesUrlFormatOfClass (in case some other class wants ValidatesUrlFormatOf::REGEXP...)
